feat:image prediction
