+++ dermysol/urls.py
@@ -18,7 +18,7 @@ from django.urls import path
 from users import views as user_views
 from django.conf import settings
 from django.conf.urls.static import static
-
+from uimage import views as uimage_views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
@@ -27,8 +27,8 @@ urlpatterns = [
     path('register/', user_views.registeruser,name='register'),
     path('login/', user_views.loginuser,name='login'),
     path('logout/', user_views.logoutuser,name='logout'),
-    path('predictimage', user_views.predictimage,name='predictimage'),
-    path('',user_views.homepage,name='home')
+    path('predictimage', uimage_views.predictimage,name='predictimage'),
+    
 
 ]
