+++ requirements.txt
@@ -19,7 +19,7 @@ numpy==1.18.5
 oauthlib==3.1.0
 opt-einsum==3.3.0
 Pillow==7.2.0
-pkg-resources==0.0.0
+
 protobuf==3.13.0
 pyasn1==0.4.8
 pyasn1-modules==0.2.8
