+++ static/css/buttonstyle.css
@@ -6,13 +6,13 @@
   border: none;
   outline: none;
   border-radius: 25px;
-  color: #fff;
+  color: #3b3a37;
   font-size: 18px;
   font-weight: 500;
   letter-spacing: 1px;
   text-transform: uppercase;
   cursor: pointer;
-  background: linear-gradient(135deg,#3a8ffe 0%,#9658fe 100%);
+  background: linear-gradient(135deg,#cceeff 0%,#00aaff 100%);
   
 }
 
@@ -24,13 +24,13 @@
   border: none;
   outline: none;
   border-radius: 25px;
-  color: #fff;
+  color:#3b3a37;
   font-size: 18px;
   font-weight: 500;
   letter-spacing: 1px;
   text-transform: uppercase;
   cursor: pointer;
-  background: linear-gradient(135deg,#3a8ffe 0%,#9658fe 100%);
+  background: linear-gradient(135deg,#cceeff 0%,#00aaff 100%);
   
 }
 
@@ -38,15 +38,15 @@
   background-color: #00b28f;
 }
 
-#custom-text {
+#imgupload .wrapper .text#custom-text{
   position: absolute;
   bottom: 0px;
   width: 100%;
   padding: 8px 0;
   font-size: 18px;
-  color: #fff;
-  display: none;
-  background: linear-gradient(135deg,#3a8ffe 0%,#9658fe 100%);
+  color: rgb(36, 34, 34);
+  display: block;
+  
 }
 @import url('https://fonts.googleapis.com/css?family=Poppins:400,500,600,700&display=swap');
 
@@ -56,19 +56,14 @@
   box-sizing: border-box;
   font-family: 'Poppins', sans-serif;
 }
-html,body{
-  display: grid;
-  height: 100%;
-  place-items: center;
-  text-align: center;
-}
 
 .container{
-  height: 350px;
-  width: 430px;
-  position: relative;
+  height: 30%;
+  width: 40%;
+  position: absolute;
+  
 }
-.container .wrapper{
+#imgupload .container .wrapper{
   position: relative;
   height: 300px;
   width: 100%;
@@ -81,10 +76,10 @@ html,body{
   overflow: hidden;
 }
 
-.wrapper.active{
+#imgupload .wrapper.active{
   border: none;
 }
-.wrapper .image{
+#imgupload .wrapper .image{
   position: absolute;
   height: 100%;
   width: 100%;
@@ -92,21 +87,25 @@ html,body{
   align-items: center;
   justify-content: center;
 }
-.wrapper .img{
+#imgupload .wrapper .img{
   height: 100%;
   width: 100%;
   object-fit: cover;
 }
-.wrapper .icon{
+#imgupload .wrapper .icon{
   font-size: 100px;
-  color: #9658fe;
+  color: #00aaff;
+  position: center;
 }
-.wrapper .text{
+/*
+#imgupload .wrapper .text{
   font-size: 20px;
   font-weight: 500;
   color: #5B5B7B;
 }
-.wrapper #cancel-btn i{
+*/
+/*
+#imgupload .wrapper #cancel-btn i{
   position: absolute;
   font-size: 20px;
   right: 15px;
@@ -121,7 +120,7 @@ html,body{
 .wrapper #cancel-btn i:hover{
   color: #e74c3c;
 }
-/*
+
 .wrapper .file-name{
   position: absolute;
   bottom: 0px;
