--- static/css/buttonstyling.css
@@ -1,124 +1,0 @@
-#custom-button {
-	margin-top: 30px;
-	display: block;
-	width: 100%;
-	height: 50px;
-	border: none;
-	outline: none;
-	border-radius: 25px;
-	color: #fff;
-	font-size: 18px;
-	font-weight: 500;
-	letter-spacing: 1px;
-	text-transform: uppercase;
-	cursor: pointer;
-	background: linear-gradient(135deg,#3a8ffe 0%,#9658fe 100%);
-	
-  }
-  
-  #submit-button {
-	margin-top: 30px;
-	display: block;
-	width: 100%;
-	height: 50px;
-	border: none;
-	outline: none;
-	border-radius: 25px;
-	color: #fff;
-	font-size: 18px;
-	font-weight: 500;
-	letter-spacing: 1px;
-	text-transform: uppercase;
-	cursor: pointer;
-	background: linear-gradient(135deg,#3a8ffe 0%,#9658fe 100%);
-	
-  }
-  
-  #custom-button:hover {
-	background-color: #00b28f;
-  }
-  
-  #custom-text {
-	position: absolute;
-	bottom: 0px;
-	width: 100%;
-	padding: 8px 0;
-	font-size: 18px;
-	color: #fff;
-	display: none;
-	background: linear-gradient(135deg,#3a8ffe 0%,#9658fe 100%);
-  }
-  @import url('https://fonts.googleapis.com/css?family=Poppins:400,500,600,700&display=swap');
-  
-  *{
-	margin: 0;
-	padding: 0;
-	box-sizing: border-box;
-	font-family: 'Poppins', sans-serif;
-  }
-  html,body{
-	display: grid;
-	height: 100%;
-	place-items: center;
-	text-align: center;
-  }
-  
-  .container {
-	height: 350px;
-	width: 430px;
-	position: relative;
-  }
-  .container .wrapper{
-	position: relative;
-	height: 300px;
-	width: 100%;
-	border-radius: 10px;
-	background: #fff;
-	border: 2px dashed #c2cdda;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	overflow: hidden;
-  }
-  
-  .wrapper.active{
-	border: none;
-  }
-  .wrapper .image{
-	position: absolute;
-	height: 100%;
-	width: 100%;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-  }
-  .wrapper .img{
-	height: 100%;
-	width: 100%;
-	object-fit: cover;
-  }
-  .wrapper .icon{
-	font-size: 100px;
-	color: #9658fe;
-  }
-  .wrapper .text{
-
-	font-size: 20px;
-	font-weight: 500;
-	color: #5B5B7B;
-  }
-  .wrapper #cancel-btn i{
-	position: absolute;
-	font-size: 20px;
-	right: 15px;
-	top: 15px;
-	color: #9658fe;
-	cursor: pointer;
-	display: none;
-  }
-  .wrapper.active:hover #cancel-btn i{
-	display: block;
-  }
-  .wrapper #cancel-btn i:hover{
-	color: #e74c3c;
-  }
