+++ static/css/displayimg.css
@@ -1,0 +1,14 @@
+body {margin:25px;}
+		
+div.polaroid {
+  width: 25%;
+  background-color: white;
+  box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+  margin-bottom: 25px;
+  position: center;
+}
+		
+div.container {
+  text-align: center;
+  padding: 10px 20px;
+}
