+++ uimage/templates/uimage/displayimage.html
@@ -4,7 +4,8 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="{% static "css/displayimg.css" %}">
+</head>
     <title>Upload</title>
     <script src="https://kit.fontawesome.com/65e590674d.js" crossorigin="anonymous"></script>
     <link rel="stylesheet" href="{% static "users/dashboard.css" %}">
@@ -22,25 +23,14 @@
 
         </div>
         <div class="content">
-           <h2>Upload Image for Classification</h2>
-            <br>
-           <form method="post" enctype="multipart/form-data" action="predictimage">
-               {% csrf_token %}
-                        <label for="fname">Upload Image File Here:</label>
-                        <div><input name="filepath" type="file" >
-                            <input type="submit" value="submit">
-                            <img src={{filePathName}} alt='imagename'>
-                            {{ filePathName }}
-                            <h2>
-                                Class : {{ prediction }}
-                            </h2>
-                    </div>
-
-
-
-           </form>
-
-
+            <div class="polaroid">
+                <img src={{filePathName}} alt='imagename' style="width: 100%;"> 
+                <br>
+                <div class="container">
+                    <h2>Class : {{ prediction }}</h2>
+                </div>
+		    </div>		           
+           
         </div>
     </div>
 </body>
