+++ uimage/templates/uimage/upload.html
@@ -1,0 +1,84 @@
+{% load static %}
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="{% static 'css/buttonstyle.css' %}">  
+    <script src="https://kit.fontawesome.com/a076d05399.js"></script>
+    <title>Upload</title>
+    <script src="https://kit.fontawesome.com/65e590674d.js" crossorigin="anonymous"></script>
+    <link rel="stylesheet" href="{% static "users/dashboard.css" %}">
+</head>
+
+<body>
+  <div class="wrapper" >
+    <div class="sidebar">
+        <h2>DERMYSOL</h2>
+        <ul>
+            <li><a href="{% url 'dashboard' %}"><i class="fas fa-columns"></i>Dashboard</a></li>
+           <li><a href="{% url 'home' %}"><i class="fas fa-home"></i>Home</a></li>
+           <li><a href="{% url 'predictimage' %}"><i class="fas fa-upload"></i>Upload</a></li>
+           <li><a href="{% url 'logout' %}"><i class="fas fa-sign-out-alt"></i>Logout</a></li>
+        </ul>
+    </div>
+    <div class="main-content">
+      <div id="imgupload">
+      <h2>Upload Image for Classification</h2>
+       <br>
+      <form method="post" enctype="multipart/form-data" action="predictimage" >
+        {% csrf_token %}
+        <div class="container" >
+          <div class="wrapper" >
+            <div class="content" >             
+            <div class="icon">
+              <i class="fas fa-cloud-upload-alt"></i>
+            </div>
+            <div class="text" id="custom-text">{{ filePathName }}</div>
+                    <!--<div class="text">No file chosen, yet!</div>-->        
+            </div>             
+          </div>
+          <div>
+            <input type="file" id="real-file" name='filepath' hidden="hidden" />
+            <button type="button" id="custom-button">CHOOSE A FILE</button>           
+          </div>
+          <div>
+            <input type="submit" id="real-submit" value="Submit" hidden="hidden" />
+            <button type="button" id="submit-button">SUBMIT FILE</button>
+          </div>
+        </div>     
+       
+              <script>
+                const wrapper = document.querySelector(".wrapper");
+                const customTxt = document.getElementById("custom-text");
+                const realFileBtn = document.getElementById("real-file");
+                const customBtn = document.getElementById("custom-button");     
+                const realSubmitBtn = document.getElementById("real-submit");
+                const customSubmit = document.getElementById("submit-button");
+                const img = document.querySelector("img");
+
+                customBtn.addEventListener("click", function() {
+                  realFileBtn.click();
+                });
+                customSubmit.addEventListener("click", function() {
+                  realSubmitBtn.click();
+                });
+                
+                realFileBtn.addEventListener("change", function() {
+                  if (realFileBtn.value) {
+                    customTxt.innerHTML = realFileBtn.value.match(
+                      /[\/\\]([\w\d\s\.\-\(\)]+)$/
+                    
+                      )[1];
+                    // img.src=realFileBtn.value;
+                  } else {
+                    customTxt.innerHTML = "No file chosen, yet.";
+                  }             
+                });
+                
+              </script>
+      </form>
+    </div>
+    </div>
+  </div>
+  </body> 
+</html>
