+++ users/static/users/dashboard.css
@@ -14,7 +14,7 @@
  .wrapper{
      display:flex;
      position: relative;
-
+     
  }
  .wrapper .sidebar{
      position: fixed;
@@ -135,13 +135,13 @@
       color:rgb(54, 58, 61);
   }
   /*uplaod*/
-  .content .form-group{
+  /*.content .form-group{
         position: relative;;
         text-align:right;
         padding:20px;
         margin:100px;
        /* background-image:url("/images/photo.png");*/
-        background-color:rgb(217, 238, 245);
+    /*    background-color:rgb(217, 238, 245);
         line-height:30px;
         border-bottom-width:20%;
 
@@ -169,3 +169,4 @@
       height:50%;
       padding:20px;
 }
+*/
